Let VSCode find lldb-vscode in current PATH.
This is achieved by using the runtime property instead of the program property.
This removes the requirement for an absolute path or the lldb-vscode binary to
be in the same directory as the extensions.
Which is probably the more common use case.

Increase the minimum required VSCode version to 1.63 (Nov 2021 Release).
Remove obsolete 'enableBreakpointsFor' configuration.
Remove unused dev dependencies and add a simple npm build script.

Add a prebuilt vsix.
Reformat package.json